% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{año_mes_dia}
\alias{año_mes_dia}
\title{Hoy}
\usage{
año_mes_dia(sep = "-")
}
\arguments{
\item{sep}{character: caracter que separa elementos; por omisión
"-"}
}
\value{
character
}
\description{
Fecha en el formato "año-mes-día"
}
\examples{
año_mes_dia() #-> 2022-09-26
}
