% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{a_propio}
\alias{a_propio}
\title{Primera mayúscula}
\usage{
a_propio(x = character())
}
\arguments{
\item{x}{character}
}
\description{
En mayúscula la primera letra de una palabra
}
\examples{
a_propio('juan calero') #-> 'Juan Calero'
}
