% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{ajustar_lon_reg_cs}
\alias{ajustar_lon_reg_cs}
\title{Cspro. Ajustar longitud de los registros}
\usage{
ajustar_lon_reg_cs(x, df_dic)
}
\arguments{
\item{x}{character: vector con los registros leídos}

\item{df_dic:}{data.frame con los datos del diccionario}
}
\value{
data.frame
}
\description{
Completar la longitud de los registros del campo
"questionnare".
}
\details{
Parece que cuando el ingreso de los datos se corta antes
del último definido en el diccionaro de datos, la cadena de
caracteres que forma un registro del campo "questionnaire"
llega hasta donde se ingresó el último. Para la lectura con
read.fwf, de los datos exportados a un archivo tipo texto, esto
es problema porque aquellas variables que están después del
último ingresado, se devuelven como character con "'NA'" (no se
"leen" porque el registro se queda corto. Puede ser que el
registro sólo tenga, digamos, 80 caracteres, y la variable
busca leerse a partir del caracter 81 o después). Además, si se
pasan argumentos al parámetro colClasses de read.fwf, se genera
un error y se detiene el proceso de lectura cuando trata de
leer datos que no existen.

\if{html}{\out{<div class="sourceCode">}}\preformatted{La función completa los registros con espacios al final,
conforme a las especificaciones en el diccionario.
}\if{html}{\out{</div>}}
}
