% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{anexar_suma_cols.data.frame}
\alias{anexar_suma_cols.data.frame}
\title{Anexar suma columnas}
\usage{
\method{anexar_suma_cols}{data.frame}(df, cols)
}
\arguments{
\item{df}{data.frame}

\item{cols}{numeric: columnas que se van a sumar. Es opcional.}
}
\value{
data.frame
}
\description{
Anexar al data.frame la suma de las columnas tipo numeric o
logical.
}
\details{
Suma todas las columnas de tipo numérico o lógico, o las
especificadas por posición (parámetro «cols»), del data.frame pasado
como argumento, y lo regresa con las sumas anexadas después de la
última fila. La suma descarta los NA. Las columnas tipo character, o
las no especificadas en «cols», son puestas a NA. Si todas las columnas
especificadas son de tipo character, devuelve el data.frame sin
modificar.
}
\examples{
\dontrun{
   x <- data.frame(x = 1:3, y = letters[1:3], z = 2.0)
   anexar_suma_cols(x)
   anexar_suma_cols(x, 3)
}
}
