% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{casar}
\alias{casar}
\title{Aparear}
\usage{
casar(..., msg = TRUE, sinpar = FALSE)
}
\arguments{
\item{...}{dos o más vectores}

\item{msg}{logical: si \code{TRUE} manda mensaje de cuántos no
hacen "match"}

\item{sinpar}{logical: si \code{TRUE} agrega el atributo "sinpar"
al resultado; \code{FALSE} por defecto.}
}
\value{
integer
}
\description{
Match de dos o más vectores
}
\details{
Para encontrar el "match" de una pareja de vectores con
otra pareja, lo común es primero construir vectores adicionales
con la función "interaction" y luego hacer el "match" con los
vectores resultantes.\preformatted{Esta función automatiza ese proceso. Recibe un número par de
vectores y construye la interacción de una mitad y de la otra
mitad y luego hace el "match". Además, de manera opcional
informa cuántos elementos de la primera mitad de vectores no
existen en la otra mitad (no hicieron "match"), y también, de
manera opcional, devuelve en el atributo "sinpar" el
correspondiente vector de índices de los que no tienen pareja.
}
}
\examples{
casar(x, y, msg = FALSE)
casar(w, x, y, z)
}
