% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{close_mysql}
\alias{close_mysql}
\title{Cierra conexión base de datos MySql}
\usage{
close_mysql(x)
}
\arguments{
\item{x}{objeto: Conexión base datos MySql}
}
\description{
Cierra conexión base de datos MySql
}
