% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{codigo_fac}
\alias{codigo_fac}
\title{fabrica código}
\usage{
codigo_fac(x, prf = "c", di = 3)
}
\arguments{
\item{x}{integer: el entero que complementa el prefijo}

\item{prf}{character: prefijo; "c" por defecto}

\item{di}{integer: dígitos que componen el código; 3 por defecto}
}
\value{
función
}
\description{
Produce una función que genera palabras con un prefijo
seguido de enteros.
}
\details{
Crea una función que produce un objeto character con
prefijo único indicado en parámetro "prf" (por omisión "c"),
seguido por un entero (parámetro "x") antecedido por
suficientes ceros para llenar tantos espacios como se indique
en el parámetro "di"
}
\examples{
ff <- codigo_fac(di = 4)
ff(4) #-> "c0004"
ff(c(20, 100)) #-> c("c0020", "c0100")
}
