% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{conn_mysql}
\alias{conn_mysql}
\title{Conectar MySQL}
\usage{
conn_mysql(...)
}
\arguments{
\item{...}{character. Argumentos a la función "dbConnect": host,
dbname, user, password.}
}
\value{
objeto DBI o NULL (si la conexión no es válida)
}
\description{
Inicia la conexión con la base de datos MySQL
}
\details{
Si alguno de los argumentos no está en ..., se toma de
"variables de ambiente"
}
\examples{
cn <- conn_mysql()
RMariaDB::dbListTables(cn)
RMariaDB::dbListFields(cn, "hato_dict")
RMariaDB::dbDisconnect(cn)
}
\seealso{
par_conn_mysql
}
