% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{conn_valido}
\alias{conn_valido}
\title{Valida conector con base de datos}
\usage{
conn_valido(cn)
}
\arguments{
\item{cn}{objeto conector con base de datos}
}
\value{
logical
}
\description{
El objeto que permite acceso a base de datos es
válido?
}
