% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{db_close.RODBC}
\alias{db_close.RODBC}
\title{Interrumpe conexión}
\usage{
\method{db_close}{RODBC}(x)
}
\arguments{
\item{x}{objeto RODBC}
}
\description{
Deshabilita el acceso a la base de datos
}
\examples{
ox <- db_xcl(system.file("demofiles/mydata.xlsx", package =
"XLConnect"))
op <- db_open(ox)
db_close(op)
}
