% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{db_columnas.RODBC}
\alias{db_columnas.RODBC}
\title{Nombre de columnas}
\usage{
\method{db_columnas}{RODBC}(x, tb = character())
}
\arguments{
\item{x}{objeto RODBC}

\item{tb}{character: nombre de la tabla}
}
\value{
character o NULL
}
\description{
Nombre de las columnas de una tabla en base de datos
}
\details{
Simplifica el resultado que devuelve la función
\code{sqlColumns} de RODBC
}
\examples{
ox <- db_xcl(system.file("demofiles/mydata.xlsx", package =
"XLConnect"))
op <- db_open(ox)
db_columnas(op, "mydata")
db_close(op)
}
