% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{db_constr.odb}
\alias{db_constr.odb}
\title{Connection string}
\usage{
\method{db_constr}{odb}(x)
}
\arguments{
\item{x}{objeto de clase odb}
}
\value{
character
}
\description{
Devuelve la 'connection string' que se genera con las
propiedades del objeto
}
\examples{
ox <- db_xcl(system.file("demofiles/mydata.xlsx", package =
"XLConnect"))
db_constr(ox)
}
