% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{db_drop.RODBC}
\alias{db_drop.RODBC}
\title{Eliminar tabla}
\usage{
\method{db_drop}{RODBC}(x, tb = character())
}
\arguments{
\item{x}{objeto RODBC}

\item{tb}{character: nombre de la tabla}
}
\value{
NULL o mensaje de error
}
\description{
elimina una tabla de la base de datos mediante
\code{sqlDrop}
}
\keyword{internal}
