% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{db_fetch.RODBC}
\alias{db_fetch.RODBC}
\title{Leer tabla}
\usage{
\method{db_fetch}{RODBC}(x, tb = character(), strfac = FALSE, max = 0L, ...)
}
\arguments{
\item{x}{objeto RODBC}

\item{tb}{character: nombre de la tabla}

\item{strfac}{logical: las cadenas de caracteres en el resultado
devueltas como factor? FALSE por defecto}

\item{max}{numeric: número máximo de registros en el resultado; 0,
sin límite, por defecto}

\item{...}{otros parámetros de \code{sqlGetResults} o
\code{sqlQuery}}
}
\value{
data.frame (invisible) o el mensaje de error
}
\description{
Los datos en una tabla de la base de datos.
}
\details{
Captura los errores y el resultado de leer todos los
campos de una tabla con \code{sqlFetch} de RODBC
}
\examples{
ox <- db_xcl(system.file("demofiles/mydata.xlsx", package =
"XLConnect"))
op <- db_open(ox)
ww <- db_fetch(op, "mydata")
db_close(op)
}
