% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{db_info.RODBC}
\alias{db_info.RODBC}
\title{Database info}
\usage{
\method{db_info}{RODBC}(x)
}
\arguments{
\item{x}{objeto RODBC}
}
\value{
character
}
\description{
Información básica sobre la base de datos y el
"driver"
}
\examples{
ox <- db_xcl(system.file("demofiles/mydata.xlsx", package =
"XLConnect"))
op <- db_open(ox)
db_info(op)
db_close(op)
}
