% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{db_open.odb}
\alias{db_open.odb}
\title{Conectar a servidor}
\usage{
\method{db_open}{odb}(x)
}
\arguments{
\item{x}{objeto odb}
}
\value{
objeto RODBC o NULL
}
\description{
Devuelve objeto que habilita llamadas al servidor de
datos.
}
\details{
Llama la función \code{odbcDriverConnect} de RODBC.
}
\examples{
ox <- db_xcl(system.file("demofiles/mydata.xlsx", package
XLConnect))
op <- db_open(ox)
}
\seealso{
\code{is_rodbc}
}
