% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{db_qry.RODBC}
\alias{db_qry.RODBC}
\title{Ejecutar consulta}
\usage{
\method{db_qry}{RODBC}(x, xs = character(), strfac = FALSE, max = 0L, ...)
}
\arguments{
\item{x}{objeto RODBC}

\item{xs}{character: expresión SQL}

\item{strfac}{logical: las cadenas de caracteres en el resultado
devueltas como factor? FALSE por defecto}

\item{max}{numeric: número máximo de registros en el resultado; 0,
sin límite, por defecto}

\item{...}{otros parámetros de \code{sqlGetResults}}
}
\value{
data.frame (invisible) o el mensaje de error
}
\description{
Devuelve resultado de una consulta a la base de datos.
}
\details{
Captura los errores y el resultado de consultar la base de
datos con la función \code{sqlQuery} de RODBC
}
\examples{
ox <- db_xcl(system.file("demofiles/mydata.xlsx", package =
"XLConnect"))
op <- db_open(ox)
ww <- db_qry(op, "select mgp, cyl from mydata")
db_close(op)
}
