% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{db_save.RODBC}
\alias{db_save.RODBC}
\title{Guardar tabla}
\usage{
\method{db_save}{RODBC}(x, df, tb = character(), rn = FALSE, ...)
}
\arguments{
\item{x}{ojeto RODBC}

\item{df}{data.frame}

\item{tb}{character: nombre de la tabla}

\item{rn}{logical: agrega nombre de filas a la tabla? FALSE por
defecto}

\item{...}{adicionales para \code{RODBC::sqlSave}}
}
\value{
NULL o mensaje de error
}
\description{
Guarda un data frame como una tabla de la base de
datos mediante \code{RODBC::sqlSave}
}
\examples{
fi <- system.file("demofiles/mydata.xlsx", package =
"XLConnect")
ox <- db_xcl(fi)
op <- db_open(ox)
ww <- data.frame(x = 1:3, y = 3:1)
##NOT RUN !!
# db_save(op, ww, "mydata")
db_close(op)
}
