% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{db_sql}
\alias{db_sql}
\title{odb SQL Server}
\usage{
db_sql(
  driver = "{SQL Server}",
  server = character(),
  database = character(),
  uid = character(),
  pwd = character()
)
}
\arguments{
\item{driver}{character: "{SQL Server}" por defecto.}

\item{server}{character: dirección IP del servidor}

\item{database}{character: nombre de la base de datos}

\item{uid}{character: nombre del usuario}

\item{pwd}{character: contraseña del usuario}
}
\value{
objeto de clase odb
}
\description{
Inicializa objeto de clase odb para conectar con un
servidor que corre SQL Server
}
\details{
La "connection string" debe contener el nombre y palabra
clave del usuario. Por seguridad, una alternativa para
"ocultar" esos datos es crear variables-ambiente definidas en
el archivo de usuario .Renviron. La función las busca en las
variables SQLUID (usuario), SQLPWD ("password") y
SQLSERVERMAGIP, si estas no se pasan entre los argumentos.
}
\examples{
ob <- db_sql(database = "encuestas")
str(ob)
}
