% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{db_tablas.RODBC}
\alias{db_tablas.RODBC}
\title{Tablas base de datos}
\usage{
\method{db_tablas}{RODBC}(x, msg = TRUE)
}
\arguments{
\item{x}{objeto RODBC}

\item{msg}{cuántas tablas? TRUE por defecto}
}
\value{
character o NULL
}
\description{
Nombres de las tablas en la base de datos.
}
\details{
Simplifica el resultado de llamar la función
\code{sqlTables} de RODBC
}
\examples{
ox <- db_xcl(system.file("demofiles/mydata.xlsx", package =
"XLConnect"))
op <- db_open(ox)
db_tablas(op)
db_close(op)
}
