% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{db_xcl}
\alias{db_xcl}
\title{odb excel}
\usage{
db_xcl(file = character(), version7 = TRUE, ronly = TRUE)
}
\arguments{
\item{file}{character: ruta de acceso al archivo}

\item{version7}{versión 2007 de excel?. TRUE por defecto.}

\item{ronly}{acceso de sólo lectura?. TRUE por defecto.}
}
\value{
objeto de clase odb
}
\description{
Inicializa objeto de clase odb para conectar con un
libro excel.
}
\examples{
ox <- db_xcl(system.file("demofiles/mydata.xlsx",
package = "XLConnect"))
str(ox)
}
