% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{decil}
\alias{decil}
\title{Decil}
\usage{
decil(x, positivos = TRUE)
}
\arguments{
\item{x}{numeric}

\item{positivos}{logical: hace el cálculo sólo con los números
mayor que 0?. TRUE por omisión.}
}
\value{
numeric
}
\description{
Deciles de una variable
}
