% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{deciles}
\alias{deciles}
\title{Deciles}
\usage{
deciles(...)
}
\arguments{
\item{...}{parámetros adicionales pasados a \code{cuantiles}}

\item{x}{numeric}
}
\value{
numeric o NA_real_
}
\description{
Deciles de una variable numérica
}
\seealso{
cuantiles
}
