% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.r
\name{df_fxp}
\alias{df_fxp}
\title{Factor data}
\usage{
df_fxp(
  dfq,
  dfp,
  qres,
  ccon = "c5000",
  cues = "quest",
  cdup = "copiade",
  cdpt = "dpt",
  cmun = "mun",
  cest = "est",
  csup = "sup",
  dft
)
}
\arguments{
\item{dfq}{data.frame con los datos de código de cuestionario,
control de llenado y control de copia}

\item{dfp}{data.frame con los datos de departamento, municipio y
estrato a los que están asignados los puntos de la muestra. Es
opcional (vea detalles).}

\item{qres}{character o numeric: códigos de control de llenado de
los cuestionarios que pueden tener datos de superficie mayor
que cero; p.ej. los con código "completo" o "incompleto"}

\item{ccon}{character: nombre de la columna con los datos de código
de control de llenado. Por omisión "c5000".}

\item{cues}{character: nombre de columna con el número de
cuestionario; por omisión, "quest"}

\item{cdup}{character: nombre de columna con el número de
cuestionario "origen". El dato debe ser NA si el cuestionario
no es "copia", o el número del cuestionario "origen" en caso
contrario. Por omisión, "copiade".}

\item{cdpt}{character: nombre de la columna con los datos de los
departamentos a los que están asignados los cuestionarios. Por
omisión "dpt".}

\item{cmun}{character: nombre de la columna con el código del
municipio al que está asignado el cuestionario. Por omisión,
"mun"}

\item{cest}{character: nombre de la columna con el código del
estrato. Por omisión, "est".}

\item{csup}{numeric: nombre de la columna con los datos de
superficie de la unidad de producción. Por omisión, "sup"}

\item{dft}{data.frame con los datos de superficie de la unidad de
producción. Es opcional (vea detalles)}
}
\value{
data.frame
}
\description{
Prepara data.frame para calcular factor de expansión
}
\details{
Prepara el data.frame con los datos necesarios para llamar
las funciones w_pr y w_nr que calculan el factor de expansión y
los ajustes por no respuesta. El data.frame que sirve de base
es el referido con el parámetro "dfq". Los arg. a los
parámetros "dft" y "dfp" son opcionales si es que arg. "dfq"
contiene todas las columnas en la lista de parámetros. Si se
pasa arg. al parámetro "dfp", este debe corresponder a los
puntos de la muestra actual; es decir, debe haberse filtrado
antes si hubiera algún esquema de rotación de la muestra.

\if{html}{\out{<div class="sourceCode">}}\preformatted{La función llama las funciones preparar_pto y preparar_sup.
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
df_fxp(dfobs, dfpun, c("completo", "incompleto"))
df_fxp(dfobs, dfpun, c(1, 3))
}
}
\seealso{
preparar_pto, preparar_sup
}
