% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.r
\name{df_pto}
\alias{df_pto}
\title{Puntos UP}
\usage{
df_pto(dfq, dfp, cues = "quest", cdpt = "dpt", cmun = "mun", cest = "estrato")
}
\arguments{
\item{dfq}{data.frame con los datos de las observaciones leídas de
la base de datos}

\item{dfp}{data.frame "maestro" con los datos de la ubicación y el
estrato al cual están asignados los puntos de la muestra actual
(la del mes).}

\item{cues}{character: columna con los números de cuestionario o
punto. El tipo de datos en la columna debe ser integer.}

\item{cdpt}{character: nombre columna con los códigos del
departamento}

\item{cmun}{character: nombre columna con el código del municipio}

\item{cest}{character: nombre columna con el código del estrato}
}
\value{
data.frame
}
\description{
Prepara el data.frame con los datos de departamento,
municipio y estrato de los puntos en la muestra
}
\details{
Hay un archivo «maestro» construido durante el diseño de
la muestra (arg. parám. "dfp") con los datos de todos los
puntos de la encuesta; este contiene: «nombre» o «id» del
punto, departamento, municipio y estrato al cual está asignado
cada punto. Si hubiera algún un esquema de rotación de la
muestra y no todos los puntos forman parte de la actual
selección, el arg. al parám. "dfp" debe haber sido previamente
filtrado para que contenga sólo los que correspondan. Esto
permite validar los datos por esa variable.

\if{html}{\out{<div class="sourceCode">}}\preformatted{Esta función lo que hace es adjuntar a los registros del
data.frame que se pasa como arg. al parám. "dfq" los datos de
departamento, municipio y estrato que están en el
arg. parám. "dfp".
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
df_pto(dfcues, dfpuntos)
}
}
