% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.r
\name{df_sup}
\alias{df_sup}
\title{Superficie UP}
\usage{
df_sup(dfq, cues = "quest", cdup = "copiade", ccon = "c5000", qres, csup, dft)
}
\arguments{
\item{dfq}{data.frame con los datos de cuestionario, de origen de
duplicados, control de cuestionario y, si acaso, superficie}

\item{cues}{character: nombre columna con los números de
cuestionario}

\item{cdup}{character: nombre columna con el origen de las
duplicadas}

\item{ccon}{character: nombre columna con el código de control}

\item{qres}{códigos de control de los cuestionarios con datos de
superficie (p.ej. de cuestionario completo o incompleto)}

\item{csup}{character: nombre columna con los datos de superficie}

\item{dft}{data.frame con los datos de cuestionario y superficie,
si es que ese dato no está en el data.frame referenciado con el
parámetro "dfq"}
}
\value{
data.frame
}
\description{
Prepara el data.frame con datos de uso de la tierra y
control de cuestionario.
}
\details{
La función produce el data.frame con los datos de
superficie y control de cuestionario, y los registros
debidamente duplicados en el caso de cuestionarios "copia". El
data.frame de cuestionarios es el referido con el parámetro
"dfq". Si este no contiene los datos de superficie, estos son
tomados del data.frame referido con el parámetro "dft", que es
opcional.
}
\examples{
\dontrun{
df_sup(x, "quest", "duplicada", "c040", "c5000", c(1, 3), y)
df_sup(x, cues = "quest", cdup = "duplicada", csup = "c040",
             ccon = "c5000", qres = c(1, 3))}
}
