% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{dia_hora_min}
\alias{dia_hora_min}
\title{Día-hora-minutos}
\usage{
dia_hora_min(sep = ".")
}
\arguments{
\item{sep}{character: separador de elementos; por omisión, "."}
}
\value{
character
}
\description{
Día, mes, año, hora, minutos en reloj del sistema
}
\examples{
dia_hora_min() #-> 26.sep.2022:19:43
}
