% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{dividir}
\alias{dividir}
\title{División}
\usage{
dividir(x = double(), y = double(), dec = NA_integer_)
}
\arguments{
\item{x}{numeric: numerador}

\item{y}{numeric: denominador}

\item{dec}{integer o NA: número de decimales del resultado; si NA
(valor por omisión), no redondea}
}
\value{
double o NA
}
\description{
Calcula el cociente y redondea.
}
\details{
Los argumentos a los parámetros deben contener el mismo
número de elementos o, si difieren, uno (numerador o
denominador) debe ser un escalar. Donde el resultado sea
infinito, devuelve NA.
}
\examples{
dividir(2, 3)
dividir(1:3, 3)
dividir(1:3, 3:1)
\dontrun{
dividir(1:3, 1:2) #-> error
}
}
\author{
eddy castellón
}
