% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{en_rango}
\alias{en_rango}
\title{Datos están en intervalo}
\usage{
en_rango(x, ref, frac = 0.1, inf, sup, inclusive = FALSE, ...)
}
\arguments{
\item{x}{numeric: los datos}

\item{ref}{numeric o function: la referencia o punto central, o la
función que lo calcula}

\item{frac}{numeric: valor mayor que cero; por omisión, 0.1}

\item{inf}{numeric: límite inferior del rango}

\item{sup}{numeric: límite superior del rango}

\item{inclusive}{logical: intervalo cerrado?; FALSE por omisión.}

\item{...}{parámetros adicionales pasados a la función indicada en «ref»}
}
\value{
logical
}
\description{
Identifica los datos comprendidos en el intervalo
delimitado por «inf» y «sup».
}
\details{
Los límites del intervalo, si no son pasados como
argumentos, se construyen multiplicando una referencia
(parámetro «ref») por un factor dependiente de una fracción
(parámetro «frac»). El factor para calcular el límite superior
es (1 + frac), y para el inferior, (1 - frac). Los datos NA se
excluyen y se devuelven tal cuales. El intervalo puede ser
cerrado o abierto, según el parámetro «inclusive»

\if{html}{\out{<div class="sourceCode">}}\preformatted{Si es omitido «ref», se asume que la función \code{mean} ocupa
su lugar
}\if{html}{\out{</div>}}
}
\examples{
x <- 1:5
en_rango(x, inf = 2, sup = 4)
en_rango(x, inf = 1, ref = 2)
en_rango(x, ref = 2, frac = 0.9)
en_rango(x, frac = 1.5)
en_rango(x, ref = median(x), frac = 1.5)
en_rango(x, ref = median, frac = 1.5, , , , na.rm = TRUE)
en_rango(x, ref = mean(x, na.rm = TRUE),
         inf = ref - 2 * sd(x, na.rm = TRUE),
         sup = ref + 2 * sd(x, na.rm = TRUE))
}
\seealso{
num_entre
}
