% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{exportar_datos_cs}
\alias{exportar_datos_cs}
\title{Exportar datos cspro}
\usage{
exportar_datos_cs(tab, dic, artx = tempfile("cs", fileext = ".txt"), cn = NULL)
}
\arguments{
\item{tab}{character: nombre de la tabla en la base de datos MySQL}

\item{dic}{data.frame: los campos del diccionario de datos}

\item{artx}{character: nombre del archivo de salida. Por omisión,
un archivo temporal con extensión ".txt" y "cs" en primeros
caracteres.}

\item{cn}{objeto para tener acceso a la base de datos. En el caso
de que sea \code{NULL} (valor por defecto) la función intenta
generar un objeto válido a partir de los parámetros de la
conexión tomados de variables de ambiente.}
}
\value{
character: nombre del archivo o NULL si no se logró
exportar
}
\description{
Exporta los datos "limpios" a un archivo tipo texto.
}
\details{
Lee los datos de la base en MySQL, ajusta la longitud de
los registros, elimina los registros marcados como borrados y
los que tienen datos duplicados en la variable "quest", bajo el
supuesto de que el último registro es el más reciente y es el
"correcto".
}
\seealso{
conn_mysql, par_con_mysql, ajustar_lon_reg_cs
}
