% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{filled_char}
\alias{filled_char}
\title{character type}
\usage{
filled_char(x)
}
\arguments{
\item{x}{vector}
}
\value{
logical
}
\description{
vector is of character type and has elements?
}
\keyword{internal}
