% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{get_atr.odb}
\alias{get_atr.odb}
\title{Get odb property}
\usage{
\method{get_atr}{odb}(x, atr)
}
\arguments{
\item{x}{objeto odb}

\item{atr}{character: atributo o propiedad}
}
\value{
valor atributo o NA
}
\description{
devuelve una propiedad
}
\examples{
ob <- odb(server = "{SQL Server}", database = "db")
get_atr(ob, "server")
}
