% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{get_dat.odb}
\alias{get_dat.odb}
\title{Datos-encuesta}
\usage{
get_dat.odb(...)
}
\arguments{
\item{...}{argumentos adicionales para la función quitar_0}

\item{x}{objeto odb}

\item{qstr}{character: expresión de consulta SQL}

\item{meta}{character: metadatos}

\item{sin_0}{logical: filtrar los registros con datos diferentes de
0? FALSE por defecto}

\item{na_0}{logical: convierte columnas numéricas de NA a cero?
FALSE por defecto}

\item{max}{numeric: número máximo de registros en el
resultado. Todos por defecto.}
}
\value{
data.frame o NULL
}
\description{
Devuelve los datos que produce una consulta SQL, y de
manera opcional, filtra los datos no igual a cero, remplaza NA
por ceros y atribuye metadatos.
}
\examples{
fi <- system.file("demofiles/multiregion.xlsx", package =
"XLConnect")
ob <- db_xcl(fi)
ww <- get_data(ob,
         "select year, days from calendar where year < 2002")
}
\keyword{internal}
