% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{guardar_excel}
\alias{guardar_excel}
\title{Guardar-excel}
\usage{
guardar_excel(
  x,
  archivo,
  hoja = "",
  sobre_hoja = FALSE,
  fila = 1L,
  columna = 1L,
  titulo = character(0),
  el_dia = TRUE
)
}
\arguments{
\item{x}{data.frame}

\item{archivo}{character: archivo excel}

\item{hoja}{character: hoja del archivo donde se va escribir. Si no
existe, se agrega al libro excel. Si no se pasa un nombre, se
agrega una nueva de nombre "Hojaxx", donde "xx" son números.}

\item{sobre_hoja}{logical: indicar que acepta escribir en una hoja
que ya existe. FALSE por omisión.}

\item{fila}{integer: número de la fila}

\item{columna}{integer: número de la columna de la celda superior
izquierda de la tabla donde se guardarán los datos}

\item{titulo}{character: una línea de título}

\item{el_dia}{logical: Escribir una celda con la fecha y hora
cuando se escribieron los datos?. TRUE por omisión}
}
\value{
nada
}
\description{
Escribir data.frame en archivo excel
}
\details{
Utiliza funciones de la librería «openxlsx» para guardar
un data.frame en archivo excel. No es permitido escribir sobre
celdas con datos.
}
