% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{ic_robusto}
\alias{ic_robusto}
\title{Intervalo robusto}
\usage{
ic_robusto(x, pctil = 2)
}
\arguments{
\item{x}{numeric: los datos}

\item{pctil}{numeric escalar: factor para determinar los
límites. Por omisión, 2.0}
}
\value{
lista con los estimados de localización "prom", dispersión
("desv") y los límites inferior y superior ("li", "ls")
}
\description{
Calcula los límites de un intervalo con estadísticas
robustas de localización y dispersión
}
\details{
Utiliza la función robustbase::covMcd para obtener los
estimados de localización y dispersión. Calcula la semiamplitud
del intervalo multiplicando el estimado de dispersión por el
factor «pctil»», y los límites superior e inferior, sumando
(restando) el estimado de localización.
}
\seealso{
en_rango
}
