% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{iniciar_vec}
\alias{iniciar_vec}
\title{Crear vector}
\usage{
iniciar_vec(busca, buscaen, remplazo, si_na = NA)
}
\arguments{
\item{busca}{character o numeric: datos a buscar}

\item{buscaen}{character o numeric: donde buscar}

\item{remplazo}{vector atómico: remplazo de los encontrados}
}
\description{
Crear vector atómico con elementos tomados de otro
vector
}
\details{
Aplica la función remplazar con el param. x igual a NULL,
para crear un vector del mismo tipo que el del param.
remplazo, y lo llena con elementos de él cuando hay un match de
los param. busca y buscaen. Si quedaran datos NA porque
elementos de "busca" no se encuentran en "buscaen", los
sustituye por el valor pasado en el parámetro si_na.
}
\examples{
iniciar_vec(c(1, 3, 2), c(2, 1), c(10, 20), -1) #-> c(20, -1, 10)
}
\seealso{
remplazar
}
