% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{is_open.RODBC}
\alias{is_open.RODBC}
\title{Canal abierto}
\usage{
\method{is_open}{RODBC}(x)
}
\arguments{
\item{x}{objeto RODBC}
}
\value{
logical
}
\description{
Habilitado el acceso a la base de datos?
}
\examples{
ox <- db_xcl(system.file("demofiles/mydata.xlsx", package =
"XLConnect"))
op <- db_open(ox)
is_open(op)
db_close(op)
}
