% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{is_rodbc}
\alias{is_rodbc}
\title{Validar RODBC}
\usage{
is_rodbc(x)
}
\arguments{
\item{x}{objeto RODBC}
}
\value{
logical
}
\description{
Es objeto RODBC?.
}
\examples{
ox <- db_xcl(system.file("demofiles/mydata.xlsx", package =
"XLConnect"))
op <- db_open(ox)
is_rodbc(op)
db_close(op)
}
