% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{leer_cuadros_fwf}
\alias{leer_cuadros_fwf}
\title{Leer y unir cuadros de datos}
\usage{
leer_cuadros_fwf(variables, columnas, tipo_col, dic, nomar)
}
\arguments{
\item{variables}{character: los nombres de las variables en el archivo}

\item{columnas}{character: nombres de las variables en el data.frame
resultante}

\item{tipo_col}{character: tipos de los datos en el archivo}

\item{dic}{data.frame: data.frame con los datos del diccionario de
datos}

\item{nomar}{character: nombre del archivo donde están todos los
datos}
}
\value{
data.frame
}
\description{
Leer datos de varias variables y unirlos en un mismo
conjunto
}
\details{
Lee los datos de varias variables relacionadas con un
mismo grupo de atributos, y los une con un rbind
}
