% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{leer_datos_fwf}
\alias{leer_datos_fwf}
\title{Leer tabla con campos adyacentes de longitud fija}
\usage{
leer_datos_fwf(
  variables = character(),
  columnas = variables,
  tipo_col = character(),
  dic,
  nomar = character()
)
}
\arguments{
\item{variables}{character: nombre de las variables en diccionario
de datos}

\item{columnas}{character: nombre de las columnas en el data.frame
que será generado. Por omisión, igual a los nombres de las
variables.}

\item{tipo_col}{character: tipo de los datos en las columnas.}

\item{dic}{data.frame: los datos del diccionario de datos}

\item{nomar}{character: nombre del archivo}
}
\value{
data.frame
}
\description{
Leer datos de un archivo compuesto de líneas de texto
con campos de longitud fija.
}
\details{
La función es una "wrap function" que utiliza la funcíón
read.fwf, adaptada para leer los datos exportados de Cspro. El
argumento al parámetro "dic" tiene las especificaciones del
diccionario de datos que determinan la posición y el número de
caracteres que ocupa cada variable en los registros.
}
\seealso{
read.fwf, exportar_datos_cs
}
