% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{longitud_variables}
\alias{longitud_variables}
\title{Cuantos caracteres ocupan las variables en registros Cspro}
\usage{
longitud_variables(x = character(), dic)
}
\arguments{
\item{x}{character: nombres de las variables}

\item{dic}{data.frame con los nombres de las variables y las
correspondientes longitudes}
}
\value{
integer
}
\description{
Construye el vector que se pasa como argumento del
parámetro "loncam" de la función get_data_cspro
}
\details{
Los elementos del vector indican el número de caracteres
que ocupa una variable dentro de la cadena de caracteres de un
registro, o bien el número de caracteres (número negativo) que
hay entre dos variables que no están contigüas en el registro.

\if{html}{\out{<div class="sourceCode">}}\preformatted{El parámetro "dic" es un data.frame con las columnas "variable"
y "length" tomadas del diccionario de datos de la tabla de
Cspro.
}\if{html}{\out{</div>}}
}
\examples{
dd <- data.frame(variable = c("boleta", "nombre", "direccion"),
                 length = c(5, 50, 100))
longitud_variables(c("boleta", "direccion"), dd) #-> c(5, -50, 100)
}
\seealso{
get_data_cspro, read.fwf
}
