% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{na0}
\alias{na0}
\title{NA a cero}
\usage{
na0(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric
}
\description{
Convierte a 0 los elementos NA de un vector de modo
numérico. Devuelve tal cual uno que no lo es.
}
\examples{
na0(c(1:3, NA_integer_))
}
