% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{na_char}
\alias{na_char}
\title{NA a character}
\usage{
na_char(x, a = "")
}
\arguments{
\item{x}{character}

\item{a}{character: caracter que sustituye a NA}
}
\value{
character
}
\description{
Convierte los NA_character_ al caracter indicado
}
