% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{nombre_propio}
\alias{nombre_propio}
\title{Nombre propio}
\usage{
nombre_propio(x = character())
}
\arguments{
\item{x}{character}
}
\value{
character
}
\description{
Ajusta las palabras a la forma de un nombre propio
}
\details{
Poda el texto, sustituye ristra de espacios y deja la
primera letra en mayúsculas
}
\examples{
a_propio('  juan   calero  ') #-> 'Juan Calero'
}
