% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{odb}
\alias{odb}
\title{Clase S3 odb}
\usage{
odb(...)
}
\arguments{
\item{...}{captura uno o más de los parámetros necesarios para
especificar la 'connection string'. Si no se indica ninguno el
resultado será una lista sin elementos, que después se podrá
llenar con la función 'setter' \code{set_odb}.}
}
\value{
objeto de clase odb
}
\description{
Crea e inicializa un objeto de clase odb.
}
\details{
Con las propiedades de este objeto se construye la cadena
de conexión ('connection string') que habilita 'conectarse' o
'abrir' la base de datos para mandar ejecutar instrucciones
de consulta.\preformatted{Para cada servidor de base de datos (SQL server, Excel, Access,
MySQL, SQLite, etc.) hay una 'connection string' específica
según la aplicación que sirve de intermediario con el servidor.

Con las librería RODBC y RODBCDBI se usan "driver" o
controladores ODBC, administrados por la aplicación
(\url{c:/windows/system32/odbcad32.exe})
"administrador de datos ODBC". El menú "controladores" muestra
los "driver" ODBC instalados en el equipo. Si no aparece el
indicado hay que descargar e instalar el "software" que lo
instala.

En \url{https://www.connectionstrings.com} y otras similares,
están descritas las 'connection string' más comunes. Estas
llevan uno o más parámetros junto con el valor que les
corresponde. Por ejemplo, la conexión ODBC mínima a libros
Excel versión 2007 requiere los parámetros 'Driver' y 'Dbq': el
primero para indicar el controlador y el segundo para la ruta
de acceso al libro: 'Driver=\{Microsoft Excel Driver (*.xls,
*.xlsx, *.xlsm, *.xlsb)\}; DBQ=\url{c:/MyExcel.xlsx};'. En este
caso, 'Driver' y 'Dbq' son las propiedades que se deben
especificar por medio de un objeto de la clase odb.
}
}
\examples{
ob <- odb(Driver = "{SQL Server}", server = "ser.o.no.ser",
          database = "dt", uid = "yo_mero", pwd = "123-probando")
}
\seealso{
\code{odb_sql}, \code{odb_xcl}
}
