% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{ok_fname}
\alias{ok_fname}
\title{file name}
\usage{
ok_fname(x = character())
}
\arguments{
\item{x}{character: nombre del archivo}
}
\value{
logical
}
\description{
Valida nombre de archivo.
}
\details{
Usa la función \code{file.create}. Para que el nombre sea
válido, cualquier directorio en la ruta de acceso, debe existir previamente.
}
\author{
eddy castellón
}
\keyword{internal}
