% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{ordenar_conforme}
\alias{ordenar_conforme}
\title{Orden conforme}
\usage{
ordenar_conforme(x, y)
}
\arguments{
\item{x}{vector atómico}

\item{y}{vector atómico que contiene a los elementos en x}
}
\value{
vector
}
\description{
Pone los elementos de un vector en el mismo orden de
los elementos de otro vector
}
\examples{
ordenar_conforme(c(2, 3, 1), c(1, 2, 3, 4)) #-> c(1, 2, 3)
ordenar_conforme(c(2, 3, 1), c(3, 1, 2, 4)) #-> c(3, 1, 2)
ordenar_conforme(c(2, 0, 1), c(1, 2, 3)) #-> c(1, 2)
}
