% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.r
\name{out_pct}
\alias{out_pct}
\title{outlier-estimación}
\usage{
out_pct(x, id, by = integer(), cota = 10L, orden = TRUE)
}
\arguments{
\item{x}{numeric: las observaciones ponderadas por el factor de
expansión.}

\item{id}{numeric o character: «id» de las observaciones}

\item{by}{numeric, character o factor: variable de agrupamiento}

\item{cota}{numeric escalar: límite superior del porcentaje de la
contribución; por omisión, 10}

\item{orden}{logical: filas del resultado en orden decreciente por
aporte al total?; por omisión TRUE}
}
\value{
data.frame o NULL, invisible
}
\description{
Identifica las observaciones cuyo aporte al estimado
del total es mayor que un límite dado.
}
\details{
Devuelve un data.frame con los datos «extremos» y el
número de observaciones involucradas en el cálculo del
total. Si no hay datos «extremos» devuelve NULL. El cálculo se
hace por grupo (departamento, por ejemplo) si así es
indicado. En este caso, el estimado del total es el estimado a
nivel de grupo.
}
\examples{
aa <- data.frame(x = 1:5, y = c(100, 1, 2, 2, 1000),
                 z = c("a", "a", "a", "b", "b"))
(out_pct(aa$y, aa$x, aa$z))
}
