% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{par_conn_mysql}
\alias{par_conn_mysql}
\title{Parámetros-conexión-RMariaDB}
\usage{
par_conn_mysql(...)
}
\arguments{
\item{...}{character: admisibles: "host", "dbname", "user", "pwd"}
}
\value{
list
}
\description{
Construye lista con los parámetros necesarios para
conectarse a una base de datos MySQL.
}
\details{
Inicia la lista con el valor de las "variables ambiente"
MYSQLSERVERMAG, MYSQLDB, MYSQLUID, MYSQLPWD, y la modifica con
los pasados como argumentos. Mensaje de alerta si alguno
"vacío".
}
\examples{
par_conn()
par_conn(dbname = "cspro")
par_conn(user = "pepe", dbname = "data", pwd = "mefistofeles")
}
