% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{parear}
\alias{parear}
\title{Aparear}
\usage{
parear(x, y)
}
\arguments{
\item{x}{character o numeric}

\item{y}{character o numeric}
}
\value{
integer
}
\description{
Alias de la función \code{match}
}
\seealso{
\code{match}, \code{casar}
}
