% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{pct}
\alias{pct}
\title{Porcentaje}
\usage{
pct(x = numeric(), base = numeric(), dec = 0L, x100 = TRUE)
}
\arguments{
\item{x}{numeric}

\item{base}{numeric: base del porcentaje; por omisión, la suma de
los datos ignorando los NA}

\item{dec}{integer: número de decimales; por omisión, cero}

\item{x100}{logical: resultado es dado multiplicado por 100 (TRUE)
o por 1 (FALSE). Por omisión, TRUE.}
}
\value{
numeric o NA
}
\description{
Porcentaje c.r.a base
}
\examples{
pct(2, 3)
pct(1:3, 3)
pct(1:3)
pct(1:3, 1:3)
\dontrun{
pct(1:3, 3:2) #-> error
}
}
\author{
eddy castellón
}
