% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{pct_grupo}
\alias{pct_grupo}
\title{Porcentaje-grupos}
\usage{
pct_grupo(x = numeric(), by = numeric(), dec = 0L)
}
\arguments{
\item{x}{numeric: datos}

\item{by}{numeric o character o factor: variable agrupamiento; por
omisión, sin agrupar}

\item{dec}{integer: decimales; 0 por omisión}
}
\value{
list
}
\description{
Contribución al total del grupo
}
\examples{
pct_grupo(1:4, c("a", "a", "b", "b"))
}
