% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{podar_str}
\alias{podar_str}
\title{Podar espacios}
\usage{
podar_str(x = character())
}
\arguments{
\item{x}{character}
}
\description{
Quita los espacios antes y después de una frase
}
\examples{
podar_str("  poda   extremos") #-> "poda   extremos"
}
