% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{print.odb}
\alias{print.odb}
\title{print.odb}
\usage{
\method{print}{odb}(x)
}
\arguments{
\item{x}{objeto odb}
}
\description{
Muestra los datos. Si existe palabra clave se muestra
como serie de "*"
}
\examples{
ob <- odb()
print(odb)
}
