% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.r
\name{puntos_estratos}
\alias{puntos_estratos}
\title{Estratos-puntos}
\usage{
puntos_estratos(es, pun)
}
\arguments{
\item{es}{data frame con los datos de los estratos}

\item{pun}{data.frame con los puntos que conforman la muestra}
}
\value{
data.frame
}
\description{
Regresa el data frame con el número actualizado de
puntos en los estratos
}
\details{
El data.frame de los estratos (pmt. "es") tiene columnas
«dpto», «estrato» y «puntos», para guardar el número de puntos
asignados a los estratos de cada departamento. El data.frame de
los puntos (pmt. "pun") tiene columnas «dpt» y «estrato», con
los códigos del departamento y del estrato al que está asignado
cada punto de la muestra. La función calcula el número de
puntos de la muestra que hay en cada estrato de cada
departamento, y lo almacena en la columna «puntos» del
data.frame de los estratos.
}
