% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{quintiles}
\alias{quintiles}
\title{Quintiles}
\usage{
quintiles(...)
}
\arguments{
\item{...}{parámetros adicionales pasados a \code{cuantiles}}

\item{x}{numeric}
}
\value{
numeric o NA_real_
}
\description{
Quintiles de una variable numérica
}
\seealso{
cuantiles
}
