% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{quitar_0.data.frame}
\alias{quitar_0.data.frame}
\title{quitar ceros}
\usage{
\method{quitar_0}{data.frame}(df, excepto = character())
}
\arguments{
\item{df}{data.frame}

\item{excepto}{nombre (character) o posición (numeric) de las
columnas de modo numeric que serán ignoradas}
}
\value{
data.frame
}
\description{
Elimina las filas de un data.frame que tienen cero en
todas las columnas de modo numeric, excepto en las columnas
indicadas.
}
\examples{
aa <- data.frame(
    x = 1:5, y = c(0, 0, 0, 1, 1),
    z = c(5, 0, 0, 2, 0)
)
(quitar_0(aa, excepto = "x"))
}
\author{
eddy castellón
}
