% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{quitar_hoja_excel}
\alias{quitar_hoja_excel}
\title{Quita hoja excel}
\usage{
quitar_hoja_excel(arch, hoja)
}
\arguments{
\item{arch}{character: nombre del archivo}

\item{hoja}{character: nombre de la hoja que se eliminará}
}
\value{
character
}
\description{
Elimina hoja de libro excel
}
