% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{read_excel_url}
\alias{read_excel_url}
\title{leer-excel-url}
\usage{
read_excel_url(url = character(), arch = character(), perm = FALSE, ...)
}
\arguments{
\item{url}{character: Dirección url}

\item{arch}{character: opcional. Ruta del archivo de descarga}

\item{perm}{logical: opcional. Conservar el archivo descargado?
(FALSE por omisión)}

\item{...}{argumentos pasados a la función read_excel}
}
\value{
data.frame o NULL
}
\description{
Leer archivo excel de acceso local o remoto
}
\details{
read_excel de readxl lee archivos locales. Si la url es
una dirección remota (un servidor externo) la función lo
descarga en el archivo indicado en el parámetro arch y luego
carga los datos deseados. Si no se indica arch, hace la
descarga a un archivo temporal. El argumento al parámetro perm,
determina si el archivo local es borrado (FALSE, opción por
defecto) o no (TRUE). En los parámetros opcionales se pasan los
argumentos necesarios para que read_excel complete la
operación.
}
\examples{
# NOT-RUN
w <- read_excel_url("https://mag.gob.ni/eddy/arch.xlsx",
                    sheet = "nov", range = "A1:B5")
w <- read_excel_url("https://mag.gob.ni/eddy/arch.xlsx",
                    arch = "c:/temp/xx.xls",
                    sheet="nov", range="A1:B5")
# lectura de archivo local
w <- read_excel_url("c:/eddy/arch.xlsx",
                    sheet="nov", range="A1:B5")
}
\seealso{
readxl::read_excel
}
