% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{redondear}
\alias{redondear}
\title{Redondear}
\usage{
redondear(x, suma = 100, metodo = "webs", q = double(), eps = 0.001)
}
\arguments{
\item{x}{numeric: números a redondear}

\item{suma}{numeric: la suma de los números redondeados (1 por
omisión)}

\item{metodo}{character: uno de "webster" (default), "adams",
"jefferson" (suficiente la primera letra)}

\item{q}{numeric: número entre 0 y 1. Es opcional. Si \code{q = 0}
es lo mismo que método "adams"; q = 1 es "jefferson"; q = 0.5
es "webster"}

\item{eps}{numeric: márgen de error (entre 0 y 1) del resultado;
por omisión, 0.001.}
}
\value{
integer
}
\description{
Redondear un vector de números de suerte que su suma
sea igual a un número dado
}
\details{
Implementa algoritmo de Dorfleitner & Klein (Statistical
Papers 40:143-157; 1999)
}
\examples{
redondear(c(1.7, 1.5, 1.0, 2.6), 8) # -> [1] 2 2 1 3
redondear(c(1.7, 1.5, 1.0, 2.6), 7) # -> [1] 2 1 1 3
redondear(c(1.7, 1.5, 1.0, 2.6), 7, q = 0.3) # -> [1] 2 1 1 3
redondear(c(1.7, 1.5, 1.0, 2.6), 7, q = 0.7) # -> [1] 2 2 1 2
}
