% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{remplazar}
\alias{remplazar}
\title{buscar-remplazar}
\usage{
remplazar(x = NULL, busca, buscaen, remplazo, msg = TRUE, toNA = TRUE)
}
\arguments{
\item{x}{vector o NULL (por omisión)}

\item{busca}{vector con los elementos a buscar}

\item{buscaen}{vector donde se buscan los elementos}

\item{remplazo}{vector con los elementos que remplazarán los
correspondientes en 'x'}

\item{msg}{TRUE por omisión; FALSE suprime mensajes de advertencia}

\item{toNA}{logical: TRUE por omisión.}
}
\value{
vector
}
\description{
Busca elementos de un vector en otro, y remplaza con
otro donde haya un match.
}
\details{
Hace un match del arg. 'busca' en el arg. 'buscaen'. Los
elementos del arg. 'remplazo' donde la función match no
devuelva NA, remplazan los correspondientes del arg. 'x'. El
número de elementos del arg. 'x' debe ser igual al del
arg. 'busca', y los del arg. 'buscaen' a los del
arg. 'remplazo'. El modo del arg. 'x' debe ser igual al de
'remplazo' (excepto cuando arg. 'x' es objeto NULL), y el modo
del arg. 'busca' al de 'buscaen'.\preformatted{El arg. 'x' es NULL por omisión. En este caso arg. 'x' se
inicializa a vector con igual número de elementos de
arg. 'busca' y mismo modo que arg. 'remplazo'. Los elementos de
arg. 'x' son ceros o NA, según lo diga el arg. 'toNA'. Son NA
si arg. 'toNA' es TRUE (por omisión).
}
}
\examples{
x <- letters[1:4]
y <- 8:1
z <- letters[1:8]
(remplazar(busca = x, buscaen = z, remplazo = y))
w <- 1:4
(remplazar(w, x, z, y))
}
\author{
eddy castellón
}
