% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{set_odb.odb}
\alias{set_odb.odb}
\title{Modificar odb}
\usage{
\method{set_odb}{odb}(x, ...)
}
\arguments{
\item{x}{objeto odb}

\item{...}{nuevas propiedades o modificaciones a las ya definidas}
}
\value{
objeto odb
}
\description{
Modifica o agrega una o más propiedades a un objeto
odb. Las propiedades previamente definidas son actualizadas a
los nuevos valores.
}
\examples{
ob <- odb(server = "{SQL Server}", database = "db")
## agrega user y modifica database
ob <- set_odb(ob, database = "ndb", user = "me")
str(ob)
}
