% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{sin_ristra_sp}
\alias{sin_ristra_sp}
\title{Sustituir espacios}
\usage{
sin_ristra_sp(x = character())
}
\arguments{
\item{x}{character}
}
\description{
Sustituye una ristra de espacios por uno solo
}
\examples{
sin_ristra_sp("   a   veces ") #-> " a veces "
}
