% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{str.odb}
\alias{str.odb}
\title{str.odb}
\usage{
\method{str}{odb}(x)
}
\arguments{
\item{x}{objeto odb}
}
\description{
Ver la estructura de objeto odb. Si hay palabra clave
se muestra como serie de "*"
}
\examples{
ob <- odb()
str(ob)
}
