% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{suma_pon}
\alias{suma_pon}
\title{Suma ponderada}
\usage{
suma_pon(x = numeric(), w = numeric())
}
\arguments{
\item{x}{numeric, logical: los datos}

\item{w}{numeric: las ponderaciones}
}
\value{
numeric
}
\description{
Suma pondera que excluye los datos y las ponderaciones NA
}
