% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{tabla_exists.RODBC}
\alias{tabla_exists.RODBC}
\title{Tabla existe}
\usage{
\method{tabla_exists}{RODBC}(x, tb = character())
}
\arguments{
\item{x}{objeto RODBC}

\item{tb}{character: nombre de tabla}
}
\value{
logical
}
\description{
existe tabla en base de datos?
}
\examples{
ox <- db_xcl(system.file("demofiles/mydata.xlsx", package =
"XLConnect"))
op <- db_open(ox)
tabla.exists(op, "mydata")
tabla.exists(op, "data")
db_close(op)
}
