% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{tok_str}
\alias{tok_str}
\title{Separar palabras}
\usage{
tok_str(str)
}
\arguments{
\item{str}{character: palabras encerradas por comillas separadas
por coma o espacios}
}
\value{
character
}
\description{
Produce un vector con las palabras (token) que se
encuentran en una ristra de caracteres, separadas unas de otras
por espacios o coma
}
\examples{
tok_str("aa bb,cc") #-> c("aa", "bb", "cc")
}
\author{
eddy castellón
}
