% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.r
\name{ubicacion_punto}
\alias{ubicacion_punto}
\title{Ubicar-punto}
\usage{
ubicacion_punto(
  punto,
  tblpun,
  colpun = "quest",
  coldpt = "dpt",
  colmun = "mun",
  colest = "estrato"
)
}
\arguments{
\item{punto}{numeric: el nombre del punto}

\item{tblpun}{data.frame: columnas con nombre del punto y códigos
de departamento, municipio y estrato}

\item{colpun}{character: nombre de la columna con «id» del punto}

\item{coldpt}{character: nombre de la columna con «id» del
departamento}

\item{colmun}{character: nombre de la columna con «id» del
municipio}

\item{colest}{character: nombre de la columna «id» del estrato}
}
\value{
data.frame con punto, departamento, municipio, estrato
}
\description{
Devuelve los códigos de departamento, municipio y
estrato donde está asignado el punto
}
\details{
Hace una búsqueda en la tabla de puntos pasada en el
parámetro tblpun.
}
