% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.r
\name{v_auto}
\alias{v_auto}
\title{autorreferencias}
\usage{
v_auto(qst, dup)
}
\arguments{
\item{qst}{numeric: códigos o «id» de los cuestionarios}

\item{dup}{numeric: códigos de los cuestionarios «origen»}
}
\value{
numeric o NULL
}
\description{
Identifica los cuestionarios «copia» con error de
autorreferencia
}
\details{
En los muestreos con reposición, la misma unidad de
muestreo puede aparecer en la muestra más de una vez; y
entonces, uno o más de los cuestionarios resulta ser «copia» de
uno que se denomina «origen». Si se quiere llevar control sobre
las «copias» para fines de manejo de los datos, por principio
todas ellas deben hacer referencia al mismo cuestionario
«origen», de modo que si hay dos o más «copias», se cae en
error de «autorreferencia» si una tiene por «origen» un
cuestionario y otra otro cuestionario «copia».\preformatted{El pmto. "qst" recibe la numeración o «id» de todos los
cuestionarios, y el pmto. "dup" la de los correspondientes
cuestionarios «origen». Si un elemento en "qst" es el «id» de
un cuestionario «copia» de otro, el correspondiente en "dup"
lleva el número o «id» de su «origen»; si no es «copia», en
"dup" va NA.

La función devuelve el «id» de los cuestionarios «copia» que
tienen como «origen» otro cuestionario «copia», o NULL si no
hay error de autorreferencia.
}
}
\examples{
aa <- 1:5
bb <- c(NA, NA, 1, 1, 3)
v_auto(aa, bb)
}
\author{
eddy castellón
}
