% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.r
\name{varianza_pps_cr}
\alias{varianza_pps_cr}
\title{varianza-pps-reemplazo}
\usage{
varianza_pps_cr(x, w = numeric(), sin0 = TRUE)
}
\arguments{
\item{x}{numeric: datos}

\item{w}{numeric: ponderación}

\item{sin0}{logical: ignorar los datos igual a cero?}
}
\value{
NA o numeric
}
\description{
Varianza pps con remplazo
}
\examples{
varianza_pps_cr(c(1, 1, 0, 1 ), c(1, 1, 1, 1 ) )
}
