% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.r
\name{w_calibra}
\alias{w_calibra}
\title{calibración}
\usage{
w_calibra(x, factor, totpob = numeric(), dec = 6L)
}
\arguments{
\item{x}{numeric: los datos de la variable por la cual se calibra}

\item{factor}{numeric: el factor de expansión}

\item{totpob}{numeric: escalar con la constante c.r.a la que se
calibra}

\item{dec}{integer: número de decimales del resultado; 6 por
omisión}
}
\value{
numeric escalar
}
\description{
Factor para ajustar factores de expansión.
}
\details{
La función calcula la división que tiene por numerador un
valor determinado (se supone que un total conocido de la
población) y por denominador la suma (descartando los NA) de
los datos (arg. "x") ponderados por las correspondientes
ponderaciones (arg. "factor"). Si los factores se multiplican
por este cociente, la suma de los datos ponderados por las
factores así ajustados, será igual (dentro de los límites de
precisión que caben) al valor utilizado como constante de
calibración.
}
\examples{
aa <- data.frame(x = 1:4, w = c(1, 1, 1, 1.5))
wc <- w_calibra(aa$x, aa$w, 10, 4)
sum(aa$x * aa$w * wc)
}
\author{
eddy castellón
}
