% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.r
\name{w_nr}
\alias{w_nr}
\title{no respuesta}
\usage{
w_nr(x, cnr, qsr, qnr, dec = 4L)
}
\arguments{
\item{x}{numérico o caracter: códigos de control del llenado del
cuestionario}

\item{cnr}{numérico o caracter: clases para agrupar los datos}

\item{qsr}{mismo tipo de x: códigos tipo "sí responde"}

\item{qnr}{mismo tipo de x: códigos tipo "rechazo" o "no responde"}

\item{dec}{integer: número de decimales al que se redondea el
resultado (4 por omisión).}
}
\value{
vector real o NULL
}
\description{
Calcula ponderación por no respuesta.
}
\details{
Es el resultado de dividir el número de unidades de
muestreo que son elegibles y que respondieron o no la
entrevista, entre el número de estas que respondieron. La
función también se puede utilizar para calcular la ponderación
por "elegilibilidad desconocida". Para este caso, el parámetro
"qnr" lleva los códigos de "elegilibilidad desconocida" (o
"no contacto") y "qsr" los de "elegilibilidad conocida"
(respondieron + no respondieron + no elegible conocido).

\if{html}{\out{<div class="sourceCode">}}\preformatted{Una unidad es "elegible" si es miembro de la población
muestreada (en una encuesta agrícola por ejemplo, las unidades
de producción no agrícola serían "no elegibles"). Es posible
que al finalizar la encuesta, la condición de elegilibilidad de
algunas unidades no sea conocida (generalmente porque no se
pudo "hacer contacto" con ella) lo que las coloca en la
categoría de "elegilibilidad desconocida".

Estas ponderaciones son una aproximación al inverso de la
probabilidad de respuesta por el que se deben multiplicar los
factores de expansión de la encuesta, a fin de reducir el sesgo
de las estimaciones debido a ese problema. La aproximación
supone que esa probabilidad es igual para todas las unidades
miembros de una misma clase (por ejemplo, todas las fincas en
un mismo municipio; o todos los productores con tales o cuales
características).
}\if{html}{\out{</div>}}
}
\examples{
clase <- c(1, 2, 1, 2, 2, 2)
ccues <- c(1, 1, 2, 1, 1, 1)
w_nr(ccues, clase, qsr = 1, qnr = 2)
}
\author{
eddy castellón
}
