% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.r
\name{w_pr}
\alias{w_pr}
\title{factores}
\usage{
w_pr(x, dfe, cob = seq.int(3), ces = seq.int(4), tamu = FALSE, dec = 6L)
}
\arguments{
\item{x}{data.frame con los datos de cuestionarios}

\item{dfe}{data.frame con los datos de los estratos}

\item{cob}{integer con los números de las columnas del data.frame x
utilizadas en el cálculo: departamento, estrato, superficie;
por omisión c(1, 2, 3).}

\item{ces}{integer con los números de las columnas del data.frame
dfe, utilizadas en el cálculo: departamento, estrato,
sup. estrato, puntos en estrato; por omisión c(1, 2, 3, 4)}

\item{tamu}{logical: calcular el número de observaciones en los
estratos y utilizarlos como el tamaño de la muestra?; FALSE por
omisión}

\item{dec}{integer: número de decimales de los resultados}
}
\value{
double o NULL
}
\description{
Calcula factor de expansión
}
\details{
El data.frame asociado al parámetro "x" tiene las columnas
que corresponden a las observaciones (cuestionarios): el
departamento (integer) y estrato (integer) a los que está
asignada la unidad de producción, y la superficie (double o
numeric). Las columnas deben aparecer en ese orden: código de
departamento seguido de el del estrato, y después la columna
con la superficie. Si no están en ese orden, se utiliza el
parámetro "cob" para indicar su posición. Por ejemplo, cob =
c(2, 5, 1) indicaría que el código de departamento está en la
segunda columna, la del estrato en la quinta, y la superficie
en la primera.

\if{html}{\out{<div class="sourceCode">}}\preformatted{El data.frame vinculado al parámetro "dfe" trae los datos
de los estratos y el tamaño de la muestra asignada a cada
uno. La posición de las columnas en el orden siguiente: la
primera con la identificación del departamento (integer),
seguida de la del estrato (integer), la superficie del estrato
(double o numeric) y el tamaño de la muestra (integer). Si no
están ese orden, se utiliza el parámetro "ces" para indicarlo,
como en el caso del data.frame de las observaciones.

El parámetro "tamu" es para calcular el factor cuando no se ha
terminado de levantar toda la muestra, considerando el número
de las presentes como el tamaño de muestra. El resultado es
estrictamente válido sólo cuando las presentes (o las que
faltan) sea una muestra aleatoria de la muestra original. La
alternativa es considerar las que faltan como unidades
"no contactadas", y ponderar el factor por no respuesta. Pero
de nuevo, el factor que resulte sería válido sólo si los
"no contactos" estuvieran distribuidos aleatoriamente.
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{w_pr(dfcues, dfestratos)}
}
