% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.r
\name{wg_calibra}
\alias{wg_calibra}
\title{calibra - grupo}
\usage{
wg_calibra(dfo, dfg, cob = 1:3, cgr = 1:2, dec = 6L)
}
\arguments{
\item{dfo}{data.frame: datos de las observaciones (vea detalles)}

\item{dfg}{data.frame: datos de los grupos (vea detalles)}

\item{cob}{numeric o character: posición (integer) o nombre
(character) de las columnas con los datos de grupo, variable de
calibración, ponderación (vea detalles); por omisión
\code{1:3}.}

\item{cgr}{numeric o character: posición (integer) o nombre
(character) de las columnas que identifican al grupo y el valor
de la variable de calibración (vea detalles); por omisión
\code{1:2}}

\item{dec}{integer: número de decimales en la ponderación; 6 por
defecto}
}
\value{
numeric
}
\description{
Calibra las ponderaciones por grupo.
}
\details{
Aplica la función \code{w_calibra} a datos agrupados,
como, por ejemplo, cuando se calibran las ponderaciones por la
superficie de los estratos (grupos) de un departamento.\preformatted{El arg. "dfo" es el data.frame con las columnas de los datos de
la variable de calibración (p.ej. superficie), del grupo al que
pertenece el dato (p.ej. estrato) y de las ponderaciones
iniciales. La posición o nombre de esas columnas se pasa en el
arg. "cob" en el orden: grupo, variable calibración y
ponderación. El arg. "dfg" es el data.frame con los datos de
los grupos: una columna con el dato que identifica al grupo
(p.ej. estrato) y otra con el valor (total) de la variable de
calibración correspondiente al grupo (p.ej. superficie del
estrato). La posición o nombre de las columnas se pasa en el
arg. "cgr" en el orden: grupo, valor de la variable.

La función devuelve el arg. "dfo" con la columna adicional «wc»
en la que están los factores de calibración.
}
}
\examples{
aa <- data.frame(x = 1:4, g = c("a", "a", "b", "b"),
                 w = c(1, 1, 1, 1.5))
bb <- data.frame(g = c("a", "b"), v = c(4, 10))
wc <- wg_calibra(aa, bb, cob = c(2, 1, 3), dec = 4)
tapply(wc$x * wc$w * wc$wc, wc$g, sum)
}
\seealso{
w_calibra
}
\author{
eddy castellón
}
