% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{xsql_st}
\alias{xsql_st}
\title{SQL expresión con o sin union}
\usage{
xsql_st(x, tabla)
}
\arguments{
\item{x}{list: con los nombres de los parámetros y los argumentos
que serán pasados a la correspondiente función. El primer
elmento, con nombre «x», indica cuál de las funciones: 1 para
llamar «xsql_s», 2 para llamar «xsql_t».}

\item{tabla}{character: nombre de la tabla de donde se extraerán
los datos}
}
\value{
character: expresión SQL
}
\description{
Llama «xsql_s» o «xsql_t» para construir la expresión
SQL. Asume que el nombre del campo que identifica los registros
es "quest".
}
\examples{
xsql_st(list(x = 1, cam = c("c5000", "copiade")),
        tabla = "seguimjul2022")
xsql_st(list(x = 2, cam = c("c001", "c002", "c003", "c004"),
        nvb = c("cul", "mz")), tabla = "seguimjul2022")
xsql_st(list(x = 2, cam = c("c001", "c002", "c003", "c004"),
        nvb = c("cul", "mz"), idc = c("mai", "fri"),
        cid = "cultivo"), tabla = "seguimjul2022")
}
\seealso{
\code{xsql_t}, \code{xsql_s}
}
