% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{ypos_si_xpos}
\alias{ypos_si_xpos}
\title{Si-entonces}
\usage{
ypos_si_xpos(x, y)
}
\arguments{
\item{x}{numeric: antecedente}

\item{y}{numeric: consecuente}
}
\value{
logical
}
\description{
Si los datos de una variable son mayor que cero, los
correspondientes de otra también lo son (condicional)
}
