% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{ypos_ssi_xpos}
\alias{ypos_ssi_xpos}
\title{equivalencia}
\usage{
ypos_ssi_xpos(x, y)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}
}
\value{
logical
}
\description{
Si los datos de una variable son mayor que cero, los
correspondientes de otra también los son, y viceversa
}
