% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.r
\name{duplicar_v}
\alias{duplicar_v}
\title{copiar vector de origen a copia}
\usage{
duplicar_v(x, qst, dup)
}
\arguments{
\item{x}{vector atómico: datos de la variable}

\item{qst}{numeric: códigos o «id» de los cuestionarios}

\item{dup}{numeric: código o «id» del cuestionario «origen» si la
boleta es «copia»; NA si no es «copia»}
}
\value{
numeric invisible
}
\description{
Copia los datos de una variable, de los cuestionarios
«origen» a los cuestionarios «copia»
}
\examples{
aa <- 1:5
bb <- 11:15
cc <- c(NA, NA, 12, NA, 11)
(duplicar_v(aa, bb, cc))
}
\author{
eddy castellón
}
