% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{exportar_datos_cs}
\alias{exportar_datos_cs}
\title{Exportar datos cspro}
\usage{
exportar_datos_cs(tab, dic, artx = tempfile("cs", fileext = ".txt"))
}
\arguments{
\item{tab}{character: nombre de la tabla en la base de datos MySQL}

\item{dic}{data.frame: los campos del diccionario de datos}

\item{artx}{character: nombre del archivo de salida. Por omisión,
un archivo temporal con extensión ".txt" y "cs" en primeros
caracteres.}
}
\value{
character: nombre del archivo
}
\description{
Exporta los datos "limpios" a un archivo tipo texto.
}
\details{
Lee los datos de la base en MySQL, ajusta la longitud de
los registros, elimina los registros marcados como borrados y
los que tienen datos duplicados en la variable "quest", bajo el
supuesto de que el último registro es el más reciente y es el
"correcto".
}
\seealso{
ajustar_lon_reg_cs
}
