% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{leer_dic_dat_xlsx}
\alias{leer_dic_dat_xlsx}
\title{Leer diccionario de datos Cspro}
\usage{
leer_dic_dat_xlsx(xlsx, filas, columnas, hoja = 1L)
}
\arguments{
\item{xlsx}{character: nombre del archivo excel}

\item{filas}{integer: número de filas que componen el diccionario}

\item{columnas}{integer: número de las columnas que conforman el
diccionario}

\item{hoja}{integer: número de la hoja que contiene el
diccionario. Por omisión, la primera hoja}
}
\value{
data.frame con los nombres de variables en minúscula.
}
\description{
Leer el diccionario de una base de datos de Cspro,
almacenado en una hoja de Excel
}
\details{
El diccionario formado por 3 columnas nombradas: variable
(llenada con los nombres de las variables del cuestionario),
start (posición del primer caracter del dato correspondiente a
la variable) y length (número de caracteres que componen el
dato)
}
\examples{
\donttest{
    leer_dic_dat_xlsx("datos/hato-dic.xlsx", 534, 2:4)
}
}
