% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut.r
\name{num_entre}
\alias{num_entre}
\title{Número-entre}
\usage{
num_entre(x, x1 = numeric(), x2 = numeric(), inclusive = FALSE)
}
\arguments{
\item{x}{numeric}

\item{x1}{numeric: límite inferior}

\item{x2}{numeric: límite superior}

\item{inclusive}{logical: con igualdad a uno de los límites?; FALSE
por omisión}
}
\value{
logical
}
\description{
Comprueba si un número está entre los límites de un
intervalo
}
\examples{
num_entre(2, 1, 2, TRUE)
num_entre(2, 1, 2)
}
