% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.r
\name{puntos_estratos}
\alias{puntos_estratos}
\title{Estratos-puntos}
\usage{
puntos_estratos(es, pun)
}
\arguments{
\item{es}{data frame con los datos de los estratos}

\item{pun}{data.frame con los puntos que conforman la muestra}
}
\value{
data.frame
}
\description{
Regresa el data frame con el número actualizado de
puntos en los estratos
}
\details{
El data.frame de los estratos tiene columnas «dpto»,
«estrato» y «puntos», que traen el número de puntos asignados a
los estratos de cada departamento. El data.frame de los puntos
tiene columnas «dpt» y «estrato» que indican el departamento y
estrato al que está asignado cada punto de la muestra. La
función modifica la columna «puntos» del data.frame de los
estratos, con el número de puntos por estrato y departamento
calculado con los datos en el data.frame de puntos de la
muestra.
}
