% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.r
\name{xsql_s}
\alias{xsql_s}
\title{SQL expresión}
\usage{
xsql_s(x = character(), cm = "*")
}
\arguments{
\item{x}{character: nombre de la tabla}

\item{cm}{character: nombre de los campos. Si los elementos con
nombre, estos son puestos en la cláusula «as». Por omisión es
"*".}
}
\value{
character
}
\description{
Expresión SQL sencilla
}
\examples{
xsql_s("a", c(x = "a", y = "b"))
#-> "select a as x,b as y from a"
xsql_s("a")
#-> "select * from a"
}
